user_repo3/requirements.txt
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-static==0.0.6
Django==1.11
django-cors-headers==2.2.0
django-extensions==2.0.7
django-filter==1.1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-firebase==0.1.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
firebase-admin==2.11.0
gcloud==0.17.0
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.12.1
gunicorn==19.8.1
httplib2==0.11.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
jws==0.1.3
Markdown==2.6.11
MarkupSafe==1.0
msgpack==0.5.6
oauth2client==3.0.0
oauthlib==2.1.0
openapi-codec==1.3.2
protobuf==3.6.0
psycopg2==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycryptodome==3.4.3
PyJWT==1.6.4
Pyrebase==3.0.27
python-firebase==1.2
python-jwt==2.0.1
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.7.0
rsa==3.4.2
simplejson==3.15.0
six==1.11.0
static3==0.7.0
uritemplate==3.0.0
urllib3==1.23
